requirements.txt
# AWS Services
boto3>=1.34.0
botocore>=1.34.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Image Processing
Pillow>=10.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.23.0

# Code Quality
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
